refactor: move util module to src/
