+++ src/main.js
@@ -2,7 +2,7 @@ import {getControl} from "./components/control";
 import {getSearch} from "./components/search";
 import {getFilters} from "./components/filter";
 import {getBoard, getBoardCards} from "./components/board";
-import {renderComponent} from "./components/util";
+import {renderComponent} from "./util";
 import {getCards, getFiltersCount} from "./data";
 
 const CARD_COUNT = 25;
